refactor: move graph into internal
