+++ packages/core/src/dsl/DeploymentBuilder.ts
@@ -28,7 +28,7 @@ import type { ModuleCache, ModuleDict, Module } from "../types/module";
 import { BigNumber, ethers } from "ethers";
 import hash from "object-hash";
 
-import { addEdge, ensureVertex } from "../graph/adjacencyList";
+import { addEdge, ensureVertex } from "../internal/graph/adjacencyList";
 import {
   CallOptions,
   ContractOptions,
