+++ packages/core/src/helpers.ts
@@ -4,4 +4,4 @@ export { serializeReplacer } from "./utils/serialize";
 export { IgnitionError, IgnitionValidationError } from "./utils/errors";
 export { TransactionsService } from "./services/TransactionsService";
 export { ContractsService } from "./services/ContractsService";
-export { VertexResultEnum } from "./types/graph";
+export { VertexResultEnum } from "./internal/types/graph";
