+++ packages/core/src/internal/execution/dispatch/executeContractDeploy.ts
@@ -7,8 +7,8 @@ import type {
 
 import { ContractFactory, ethers } from "ethers";
 
-import { VertexResultEnum } from "../../../types/graph";
 import { collectLibrariesAndLink } from "../../../utils/collectLibrariesAndLink";
+import { VertexResultEnum } from "../../types/graph";
 
 import { resolveFrom, toAddress } from "./utils";
