+++ packages/core/src/internal/execution/dispatch/executeDeployedContract.ts
@@ -5,7 +5,7 @@ import type {
   ExecutionVertexVisitResult,
 } from "../../types/executionGraph";
 
-import { VertexResultEnum } from "../../../types/graph";
+import { VertexResultEnum } from "../../types/graph";
 
 import { resolveFrom, toAddress } from "./utils";
