+++ packages/core/src/internal/execution/dispatch/executeSendETH.ts
@@ -5,7 +5,7 @@ import type {
 } from "../../types/executionGraph";
 import type { PopulatedTransaction } from "ethers";
 
-import { VertexResultEnum } from "../../../types/graph";
+import { VertexResultEnum } from "../../types/graph";
 
 import { resolveFrom, toAddress } from "./utils";
