+++ packages/core/src/internal/execution/utils.ts
@@ -1,6 +1,6 @@
 import hash from "object-hash";
 
-import { getDependenciesFor } from "../../graph/adjacencyList";
+import { getDependenciesFor } from "../../internal/graph/adjacencyList";
 import { serializeReplacer } from "../../utils/serialize";
 import { ExecutionGraph } from "../execution/ExecutionGraph";
