+++ packages/core/src/internal/graph/utils.ts
