+++ packages/core/src/internal/types/graph.ts
