+++ packages/core/src/validation/dispatch/validateArtifactContract.ts
@@ -1,7 +1,7 @@
 import { ethers, BigNumber } from "ethers";
 
 import { ArtifactContractDeploymentVertex } from "../../internal/types/deploymentGraph";
-import { VertexResultEnum } from "../../types/graph";
+import { VertexResultEnum } from "../../internal/types/graph";
 import {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
