+++ packages/core/src/validation/dispatch/validateHardhatContract.ts
@@ -1,7 +1,7 @@
 import { ethers, BigNumber } from "ethers";
 
 import { HardhatContractDeploymentVertex } from "../../internal/types/deploymentGraph";
-import { VertexResultEnum } from "../../types/graph";
+import { VertexResultEnum } from "../../internal/types/graph";
 import {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
