+++ packages/core/src/validation/dispatch/validateSendETH.ts
@@ -1,7 +1,7 @@
 import { ethers, BigNumber } from "ethers";
 
 import { SendVertex } from "../../internal/types/deploymentGraph";
-import { VertexResultEnum } from "../../types/graph";
+import { VertexResultEnum } from "../../internal/types/graph";
 import {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
