+++ packages/core/test/execution/batching.ts
@@ -11,7 +11,7 @@ import { BigNumber } from "ethers";
 import { Deployment } from "../../src/internal/deployment/Deployment";
 import { ExecutionGraph } from "../../src/internal/execution/ExecutionGraph";
 import { executeInBatches } from "../../src/internal/execution/execute";
-import { VertexResultEnum } from "../../src/types/graph";
+import { VertexResultEnum } from "../../src/internal/types/graph";
 import { ICommandJournal } from "../../src/types/journal";
 import { buildAdjacencyListFrom } from "../graph/helpers";
